Override dynamic plugin sdk module in runtime
