+++ frontend/@types/console/index.d.ts
@@ -55,8 +55,6 @@ declare interface Window {
   pluginStore?: {}; // Console plugin store
   loadPluginEntry?: Function;
   Cypress?: {};
-  api: {}; // Console dynamic plugin APIs
-  internalAPI: {};
 }
 
 // TODO: Remove when upgrading to TypeScript 4.1.2+, which has a type for RelativeTimeFormat.
