+++ frontend/packages/console-dynamic-plugin-sdk/scripts/generate-pkg-assets.ts
@@ -9,13 +9,7 @@ const createPackageJson = (packagePath: string) => {
   packageJson.name = '@openshift-console/dynamic-plugin-sdk';
   delete packageJson.private;
   packageJson.license = 'Apache-2.0';
-  packageJson.main = 'lib/extensions/index.js';
-  packageJson.exports = {
-    '.': './lib/extensions/index.js',
-    './webpack': './lib/webpack/ConsoleRemotePlugin.js',
-    './api': './lib/api/api.js',
-    './internalAPI': './lib/api/internal-api.js',
-  };
+  packageJson.main = 'lib/index-lib.js';
   packageJson.readme = 'README.md';
   packageJson.peerDependencies = _.pick(packageJson.devDependencies, 'webpack');
   delete packageJson.dependencies;
