--- frontend/packages/console-dynamic-plugin-sdk/src/api/component-api-types.ts
@@ -1,15 +1,0 @@
-import { RouteComponentProps } from 'react-router';
-import { K8sResourceCommon } from '../extensions';
-
-/* Horizontal Nav Types */
-export type NavPage = {
-  href?: string;
-  path?: string;
-  name: string;
-  component: React.ComponentType<RouteComponentProps>;
-};
-
-export type HorizontalNavProps = {
-  resource?: K8sResourceCommon;
-  pages: NavPage[];
-};
