--- frontend/packages/console-dynamic-plugin-sdk/src/api/internal/index.ts
@@ -1,2 +1,0 @@
-export * from './dashboard/types';
-export * from './dashboard/api';
