+++ frontend/packages/console-dynamic-plugin-sdk/src/extensions/storage-provider.ts
@@ -1,6 +1,5 @@
 import { RouteComponentProps } from 'react-router';
-import { Extension } from '@console/plugin-sdk/src/typings/base';
-import { ExtensionDeclaration, CodeRef } from '../types';
+import { Extension, ExtensionDeclaration, CodeRef } from '../types';
 
 export type StorageProvider = ExtensionDeclaration<
   'console.storage-provider',
