+++ frontend/packages/console-dynamic-plugin-sdk/src/extensions/topology-details.ts
@@ -1,9 +1,8 @@
 import * as React from 'react';
 import { GraphElement } from '@patternfly/react-topology';
-import { Extension } from '@console/plugin-sdk/src/typings/base';
-import { BuildConfigData } from '../api/topology-types';
-import { CodeRef, ExtensionDeclaration } from '../types';
+import { Extension, CodeRef, ExtensionDeclaration } from '../types';
 import { K8sResourceCommon } from './console-types';
+import { BuildConfigData } from './topology-types';
 
 /** DetailsTab contributes a tab for the topology details panel. */
 export type DetailsTab = ExtensionDeclaration<
