+++ frontend/packages/console-dynamic-plugin-sdk/src/index-lib.ts
@@ -1,0 +1,2 @@
+export * from './extensions';
+export * from './api/api';
