+++ frontend/packages/console-dynamic-plugin-sdk/src/index-webpack.ts
@@ -1,0 +1,1 @@
+export { ConsoleRemotePlugin } from './webpack/ConsoleRemotePlugin';
