+++ frontend/packages/console-shared/src/components/dashboard/details-card/DetailsBody.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import './details-card.scss';
-import { DetailsBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal';
+import { DetailsBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 
 const DetailsBody: React.FC<DetailsBodyProps> = ({ children }) => (
   <dl className="co-dashboard-card__body--top-margin co-details-card__body co-dashboard-text--small">
