+++ frontend/packages/console-shared/src/components/dashboard/status-card/AlertsBody.tsx
@@ -7,7 +7,7 @@ import {
 } from '@patternfly/react-core';
 import { UnknownIcon } from '@patternfly/react-icons';
 import { useTranslation } from 'react-i18next';
-import { AlertsBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal';
+import { AlertsBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 
 const AlertsBody: React.FC<AlertsBodyProps> = ({ error = false, children }) => {
   const { t } = useTranslation();
