+++ frontend/packages/console-shared/src/components/dashboard/utilization-card/UtilizationBody.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { ChartAxis, ChartContainer } from '@patternfly/react-charts';
 import { Grid } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
-import { UtilizationBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal';
+import { UtilizationBodyProps } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { timeFormatter } from '@console/internal/components/utils/datetime';
 import { useRefWidth } from '@console/internal/components/utils/ref-width-hook';
 import { useUtilizationDuration } from '@console/shared';
