+++ frontend/packages/console-shared/src/components/dashboard/utilization-card/UtilizationDurationDropdown.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import { UtilizationDurationDropdownProps } from '@console/dynamic-plugin-sdk/src/api/internal';
+import { UtilizationDurationDropdownProps } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { Dropdown } from '@console/internal/components/utils';
 import { DurationKeys, DURATION_VALUES, useUtilizationDuration } from '@console/shared';
