+++ frontend/packages/console-shared/src/graph-helper/data-utils.ts
@@ -1,5 +1,5 @@
 import * as _ from 'lodash';
-import { ByteDataTypes } from '@console/dynamic-plugin-sdk/src/api/internal';
+import { ByteDataTypes } from '@console/dynamic-plugin-sdk/src/api/internal-types';
 import { DataPoint } from '@console/internal/components/graphs';
 import { getType } from '@console/internal/components/utils/units';
