+++ frontend/public/components/utils/k8s-watch-hook.ts
@@ -7,10 +7,10 @@ import { createSelectorCreator, defaultMemoize } from 'reselect';
 import { useDeepCompareMemoize } from '@console/shared/src/hooks/deep-compare-memoize';
 import { usePrevious } from '@console/shared/src/hooks/previous';
 import {
+  WatchK8sResource,
   UseK8sWatchResource,
   UseK8sWatchResources,
-  WatchK8sResource,
-} from '@console/dynamic-plugin-sdk/src/api/api-types';
+} from '@console/dynamic-plugin-sdk/src/extensions/console-types';
 
 import { makeQuery, makeReduxID } from './k8s-watcher';
 import * as k8sActions from '../../actions/k8s';
@@ -19,14 +19,15 @@ import { RootState } from '../../redux';
 import { K8sState } from '../../reducers/k8s';
 
 export {
-  UseK8sWatchResource,
-  WatchK8sResult,
   ResourcesObject,
   WatchK8sResults,
   WatchK8sResultsObject,
   WatchK8sResource,
   WatchK8sResources,
-} from '@console/dynamic-plugin-sdk/src/api/api-types';
+  UseK8sWatchResource,
+  UseK8sWatchResources,
+  WatchK8sResult,
+} from '@console/dynamic-plugin-sdk/src/extensions/console-types';
 
 export class NoModelError extends Error {
   constructor() {
